% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavisualizationpackager.R
\docType{package}
\name{datavisualizationpackager}
\alias{datavisualizationpackager-package}
\alias{datavisualizationpackager}
\title{datavisualizationpackager: A Package for three different datavisualization}
\description{
This datavisualization package contains three distinct functions for data visualization:
animated_lineplot: Generates an animated line plot of your data,
boxplot_with_mean: Creates a boxplot overlaid with a line representing the mean of your data,
scatterplot_with_regression: Produces a scatterplot with the corresponding regression line of your data.
}
\section{Functions}{

\itemize{
\item \code{\link{animated_lineplot}}: Generates an animated line plot of your data.
\item \code{\link{boxplot_with_mean}}: Creates a boxplot overlaid with a line representing the mean of your data.
\item \code{\link{scatterplot_with_regression}}: Produces a scatterplot with the corresponding regression line of your data.
}
}

\keyword{internal}
