% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavisualizationpackager.R
\name{scatterplot_with_regression}
\alias{scatterplot_with_regression}
\title{Scatterplot with regressionline}
\usage{
scatterplot_with_regression(
  data,
  x,
  y,
  main = "Scatterplot with Regression Line"
)
}
\arguments{
\item{data}{A dataframe with the data to be visualized.}

\item{x}{The variable plotted on the x-axis (as a string).}

\item{y}{The variable plotted on the y-axis (as a string).}

\item{main}{(optional) The title of the plot. Defaults to "Scatterplot with Regression Line".}
}
\description{
Creates a scatterplot of the given data and adds a regression line.
}
\examples{
# Example dataset
set.seed(123)
data <- data.frame(x = rnorm(100), y = rnorm(100))

# Create a scatterplot with a regression line
scatterplot_with_regression(data, x = "x", y = "y", main = "Example Scatterplot with Regression Line")
}
